replace internal traits instead of removing
removing traits is not guaranteed without a scheduled task to ensure the CitizensTraitAdapter is run last. Scheduled task may still allow the trait to partially execute.

The replacement traits extend the trait they are replacing and prevent the trait from executing when attached to a provider based NPC. The trait runs normally when attached to a Citizens NPC.